feat(data-table): Now we can use this component to quickly generate a table
