+++ src/components/data-table/table-loading.vue
@@ -1,0 +1,13 @@
+<script lang="ts" setup>
+import { Loader } from 'lucide-vue-next'
+</script>
+
+<template>
+  <div class="h-120 w-full  flex items-center justify-center">
+    <Loader class="animate-spin " />
+  </div>
+</template>
+
+<style scoped>
+
+</style>
