+++ src/components/data-table/use-generate-vue-table.ts
@@ -1,0 +1,37 @@
+import type { ColumnFiltersState, SortingState, VisibilityState } from '@tanstack/vue-table'
+
+import { getCoreRowModel, getFacetedRowModel, getFacetedUniqueValues, getFilteredRowModel, getPaginationRowModel, getSortedRowModel, useVueTable } from '@tanstack/vue-table'
+
+import { valueUpdater } from '@/lib/utils'
+
+import type { DataTableProps } from './types'
+
+export function generateVueTable<T>(props: DataTableProps<T>) {
+  const sorting = ref<SortingState>([])
+  const columnFilters = ref<ColumnFiltersState>([])
+  const columnVisibility = ref<VisibilityState>({})
+  const rowSelection = ref({})
+
+  const table = useVueTable({
+    get data() { return props.data },
+    get columns() { return props.columns },
+    state: {
+      get sorting() { return sorting.value },
+      get columnFilters() { return columnFilters.value },
+      get columnVisibility() { return columnVisibility.value },
+      get rowSelection() { return rowSelection.value },
+    },
+    enableRowSelection: true,
+    onSortingChange: updaterOrValue => valueUpdater(updaterOrValue, sorting),
+    onColumnFiltersChange: updaterOrValue => valueUpdater(updaterOrValue, columnFilters),
+    onColumnVisibilityChange: updaterOrValue => valueUpdater(updaterOrValue, columnVisibility),
+    onRowSelectionChange: updaterOrValue => valueUpdater(updaterOrValue, rowSelection),
+    getCoreRowModel: getCoreRowModel(),
+    getFilteredRowModel: getFilteredRowModel(),
+    getPaginationRowModel: getPaginationRowModel(),
+    getSortedRowModel: getSortedRowModel(),
+    getFacetedRowModel: getFacetedRowModel(),
+    getFacetedUniqueValues: getFacetedUniqueValues(),
+  })
+  return table
+}
