+++ src/types/auto-import-components.d.ts
@@ -23,8 +23,10 @@ declare module 'vue' {
     CustomThemeThemePopover: typeof import('./../components/custom-theme/theme-popover.vue')['default']
     CustomThemeTitle: typeof import('./../components/custom-theme/custom-theme-title.vue')['default']
     CustomThemeToggleColorMode: typeof import('./../components/custom-theme/toggle-color-mode.vue')['default']
+    DataTable: typeof import('./../components/data-table/data-table.vue')['default']
     DataTableColumnHeader: typeof import('./../components/data-table/column-header.vue')['default']
     DataTableFacetedFilter: typeof import('./../components/data-table/faceted-filter.vue')['default']
+    DataTableTableLoading: typeof import('./../components/data-table/table-loading.vue')['default']
     DataTableTablePagination: typeof import('./../components/data-table/table-pagination.vue')['default']
     DataTableViewOptions: typeof import('./../components/data-table/view-options.vue')['default']
     GlobalLayoutBasicHeader: typeof import('./../components/global-layout/basic-header.vue')['default']
